enhancements, added daysThreshold param
