+++ src/main/resources/application.properties
@@ -66,5 +66,6 @@ advertisement.images.big.width=800
 advertisement.images.big.height=800
 advertisement.images.small.width=300
 advertisement.images.small.height=300
+advertisement.draft.images.daysThreshold=10
 users.images.avatar.width=300
 users.images.avatar.height=300
